# Conflicts:
#	.env
#	megano/megano/settings.py

+++ megano/auth_app/repositories/auth_repositories.py
@@ -20,14 +20,17 @@ class AuthRepository(IAuth):
             return None
 
     @beartype
-    def delete_user_by_email(self, _email: str) -> Optional[bool]:
+    def get_user(self, model: User) -> User:
+        """Получаем пользователя"""
+        user = model
+        return user
+
+    @beartype
+    def delete_user_by_email(self, _email: str) -> None:
         """Удаляем пользователя"""
-        try:
-            user = self.get_user_by_email(_email)
+        user = self.get_user_by_email(_email)
+        if user:
             user.delete()
-            return True
-        except ObjectDoesNotExist:
-            return None
 
     @beartype
     def get_user_by_activation_key(self, _activation_key: str) -> Optional[User]:
@@ -36,8 +39,3 @@ class AuthRepository(IAuth):
             return User.objects.get(activation_key=_activation_key)
         except ObjectDoesNotExist:
             return None
-
-    @beartype
-    def set_user_is_active(self, model: User, value: bool) -> None:
-        """Установить значение параметра is_active пользователя"""
-        model.is_active = value
